Move twitter module to its own package (#490)
