+++ cmd/bot/minit.go
@@ -16,4 +16,6 @@ import (
 	_ "github.com/pajbot/pajbot2/pkg/modules/link_filter"
 
 	_ "github.com/pajbot/pajbot2/pkg/modules/nuke"
+
+	_ "github.com/pajbot/pajbot2/pkg/modules/twitter"
 )
